requirements.txt
flask
gunicorn
python-dotenv
celery
sqlalchemy
supervisor